ISIS-2726: merges QueryResultsCache and QueryResultsCacheDefault
