+++ api/applib/src/main/java/org/apache/isis/applib/services/queryresultscache/QueryResultsCache.java
@@ -19,17 +19,34 @@
 package org.apache.isis.applib.services.queryresultscache;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.Callable;
 
+import javax.inject.Named;
+
 import org.springframework.beans.factory.DisposableBean;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Primary;
+import org.springframework.core.annotation.Order;
+import org.springframework.stereotype.Service;
+
+import org.apache.isis.applib.annotation.InteractionScope;
+import org.apache.isis.applib.annotation.OrderPrecedence;
+import org.apache.isis.commons.internal.base._Casts;
+import org.apache.isis.commons.internal.base._NullSafe;
+import org.apache.isis.commons.internal.collections._Maps;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.extern.log4j.Log4j2;
 
 
 /**
- * Provides a mechanism by which idempotent query results can be cached for
- * the duration of an interaction.
+ * This service (API and implementation) provides a mechanism by which
+ * idempotent query results can be cached for the duration of an {@link org.apache.isis.applib.services.iactn.Interaction}.
  *
  * <p>
  * Caching such values is useful to improve the response time (for the end user)
@@ -38,9 +55,24 @@ import lombok.Getter;
  * expensive operation is made many times, then the query cache is a perfect fit.
  * </p>
  *
+ * <p>
+ * For example, code that makes a repository call many times within a loop
+ * can be performance tuned using this service.  The benefit is that the
+ * algorithm of the business logic can remain easy to understand.
+ * </p>
+ *
  * @since 1.x {@index}
  */
-public interface QueryResultsCache extends DisposableBean {
+@Service
+@Named("isis.applib.QueryResultsCache")
+@Order(OrderPrecedence.EARLY)
+@Primary
+@InteractionScope
+@Qualifier("Default")
+@Log4j2
+public class QueryResultsCache implements DisposableBean {
+
+    private final Map<Key, Value<?>> cache = _Maps.newHashMap();
 
     /**
      * Executes the callable if not already cached for the supplied calling
@@ -52,101 +84,90 @@ public interface QueryResultsCache extends DisposableBean {
      * @param keys
      * @param <T>
      */
-    <T> T execute(
-            Callable<T> callable,
-            Class<?> callingClass, String methodName,
-            Object... keys);
+    public <T> T execute(
+            final Callable<T> callable,
+            final Class<?> callingClass,
+            final String methodName,
+            final Object... keys) {
+        if(isIgnoreCache()) {
+            try {
+                return callable.call();
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+        final Key cacheKey = new Key(callingClass, methodName, keys);
+        return executeWithCaching(callable, cacheKey);
+    }
 
 
-    /**
-     * Not API, for framework use only.
-     */
-    void onTransactionEnded();
 
-    <R> R execute(
-            MethodReferences.Call0<? extends R> action,
-            Class<?> callingClass, String methodName);
 
-    <R, A0> R execute(
-            MethodReferences.Call1<? extends R, A0> action,
-            Class<?> callingClass, String methodName,
-            A0 arg0);
+    public <R> R execute(MethodReferences.Call0<? extends R> action, Class<?> callingClass, String methodName) {
+        if(isIgnoreCache()) {
+            return action.call();
+        }
+        final Key cacheKey = new Key(callingClass, methodName);
+        return executeWithCaching(action::call, cacheKey);
+    }
 
-    <R, A0, A1> R execute(
-            MethodReferences.Call2<? extends R, A0, A1> action,
-            Class<?> callingClass, String methodName,
-            A0 arg0, A1 arg1);
 
-    <R, A0, A1, A2> R execute(
-            MethodReferences.Call3<? extends R, A0, A1, A2> action,
-            Class<?> callingClass, String methodName,
-            A0 arg0, A1 arg1, A2 arg2);
+    public <R, A0> R execute(MethodReferences.Call1<? extends R, A0> action, Class<?> callingClass, String methodName, A0 arg0) {
+        if(isIgnoreCache()) {
+            return action.call(arg0);
+        }
+        final Key cacheKey = new Key(callingClass, methodName, arg0);
+        return executeWithCaching(()->action.call(arg0), cacheKey);
+    }
 
-    <R, A0, A1, A2, A3> R execute(
-            MethodReferences.Call4<? extends R, A0, A1, A2, A3> action,
-            Class<?> callingClass, String methodName,
-            A0 arg0, A1 arg1, A2 arg2, A3 arg3);
+    public <R, A0, A1> R execute(MethodReferences.Call2<? extends R, A0, A1> action, Class<?> callingClass, String methodName, A0 arg0,
+                                 A1 arg1) {
+        if(isIgnoreCache()) {
+            return action.call(arg0, arg1);
+        }
+        final Key cacheKey = new Key(callingClass, methodName, arg0, arg1);
+        return executeWithCaching(()->action.call(arg0, arg1), cacheKey);
+    }
 
-    <R, A0, A1, A2, A3, A4> R execute(
-            MethodReferences.Call5<? extends R, A0, A1, A2, A3, A4> action,
-            Class<?> callingClass, String methodName,
-            A0 arg0, A1 arg1, A2 arg2, A3 arg3, A4 arg4);
+    public <R, A0, A1, A2> R execute(MethodReferences.Call3<? extends R, A0, A1, A2> action, Class<?> callingClass, String methodName,
+                                     A0 arg0, A1 arg1, A2 arg2) {
+        if(isIgnoreCache()) {
+            return action.call(arg0, arg1, arg2);
+        }
+        final Key cacheKey = new Key(callingClass, methodName, arg0, arg1, arg2);
+        return executeWithCaching(()->action.call(arg0, arg1, arg2), cacheKey);
+    }
 
-    class Key {
+    public <R, A0, A1, A2, A3> R execute(MethodReferences.Call4<? extends R, A0, A1, A2, A3> action, Class<?> callingClass,
+                                         String methodName, A0 arg0, A1 arg1, A2 arg2, A3 arg3) {
+        if(isIgnoreCache()) {
+            return action.call(arg0, arg1, arg2, arg3);
+        }
+        final Key cacheKey = new Key(callingClass, methodName, arg0, arg1, arg2, arg3);
+        return executeWithCaching(()->action.call(arg0, arg1, arg2, arg3), cacheKey);
+    }
 
-        @Getter
-        private final Class<?> callingClass;
-        @Getter
-        private final String methodName;
-        @Getter
-        private final Object[] keys;
+    public <R, A0, A1, A2, A3, A4> R execute(MethodReferences.Call5<? extends R, A0, A1, A2, A3, A4> action, Class<?> callingClass,
+                                             String methodName, A0 arg0, A1 arg1, A2 arg2, A3 arg3, A4 arg4) {
+        if(isIgnoreCache()) {
+            return action.call(arg0, arg1, arg2, arg3, arg4);
+        }
+        final Key cacheKey = new Key(callingClass, methodName, arg0, arg1, arg2, arg3, arg4);
+        return executeWithCaching(()->action.call(arg0, arg1, arg2, arg3, arg4), cacheKey);
+    }
 
+    @Getter @EqualsAndHashCode
+    public static class Key {
+
+        @Getter private final Class<?> callingClass;
+        @Getter private final String methodName;
+        @Getter private final Object[] keys;
+
+        // not using @RequiredArgsConstructor as we have used varargs here
         public Key(Class<?> callingClass, String methodName, Object... keys) {
             this.callingClass = callingClass;
             this.methodName = methodName;
             this.keys = keys;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj)
-                return true;
-            if (obj == null)
-                return false;
-            if (getClass() != obj.getClass())
-                return false;
-            Key other = (Key) obj;
-
-            // compare callingClass
-            if (callingClass == null) {
-                if (other.callingClass != null)
-                    return false;
-            } else if (!callingClass.equals(other.callingClass))
-                return false;
-
-            // compare methodName
-            if (methodName == null) {
-                if (other.methodName != null)
-                    return false;
-            } else if (!methodName.equals(other.methodName))
-                return false;
-
-            // compare keys
-            if (!Arrays.equals(keys, other.keys))
-                return false;
-
-            // ok, matches
-            return true;
-        }
-
-        @Override
-        public int hashCode() {
-            final int prime = 31;
-            int result = 1;
-            result = prime * result + ((callingClass == null) ? 0 : callingClass.hashCode());
-            result = prime * result + Arrays.hashCode(keys);
-            result = prime * result + ((methodName == null) ? 0 : methodName.hashCode());
-            return result;
         }
 
         @Override
@@ -158,6 +179,74 @@ public interface QueryResultsCache extends DisposableBean {
     @Data
     class Value<T> {
         private final T result;
+    }
+
+
+
+    private <T> T executeWithCaching(final Callable<T> callable, final Key cacheKey) {
+        try {
+            final Value<?> cacheValue = cache.get(cacheKey);
+            logHitOrMiss(cacheKey, cacheValue);
+            if(cacheValue != null) {
+                return _Casts.uncheckedCast(cacheValue.getResult());
+            }
+
+            // cache miss, so get the result...
+            T result = callable.call();
+
+            // ... and cache
+            //
+            // (it is possible that the callable just invoked might also have updated the cache, eg if there was
+            // some sort of recursion.  However, Map#put(...) is idempotent, so valid to call more than once.
+            //
+            // note: there's no need for thread-safety synchronization... remember that QueryResultsCache is @RequestScoped
+            put(cacheKey, result);
+
+            return result;
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private <T> void put(final Key cacheKey, final T result) {
+        log.debug("PUT: {}", cacheKey);
+        cache.put(cacheKey, new Value<T>(result));
+    }
+
+    private static void logHitOrMiss(final Key cacheKey, final Value<?> cacheValue) {
+        if(!log.isDebugEnabled()) {
+            return;
+        }
+        log.debug("{}: {}", (cacheValue != null ? "HIT" : "MISS"), cacheKey.toString());
+    }
+
+
+    /**
+     * Not API: for framework to call at end of transaction, to clear out the cache.
+     *
+     * <p>
+     * (This service really ought to be considered
+     * a transaction-scoped service; since that isn't yet supported by the framework, we have to manually reset).
+     * </p>
+     */
+    public void onTransactionEnded() {
+        cache.clear();
+    }
+
+
+    @Override
+    public void destroy() throws Exception {
+        cache.clear();
+    }
+
+    // -- HELPER
+
+    @Autowired(required = false)
+    protected List<QueryResultsCacheControl> cacheControl;
+
+    private boolean isIgnoreCache() {
+        return _NullSafe.stream(cacheControl)
+                .anyMatch(c->c.isIgnoreCache());
     }
 
 }
