+++ core/config/src/main/java/org/apache/isis/core/config/presets/DebugTransactionScopedServices.properties
@@ -5,9 +5,9 @@
 #  to you under the Apache License, Version 2.0 (the
 #  "License"); you may not use this file except in compliance
 #  with the License.  You may obtain a copy of the License at
-#  
+#
 #         http://www.apache.org/licenses/LICENSE-2.0
-#         
+#
 #  Unless required by applicable law or agreed to in writing,
 #  software distributed under the License is distributed on an
 #  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -17,5 +17,5 @@
 
 # -- framework internal transaction scoped --
 logging.level.org.apache.isis.persistence.jdo.datanucleus5.metrics.MetricsServiceForJdo = DEBUG
-logging.level.org.apache.isis.core.runtimeservices.queryresultscache.QueryResultsCacheDefault = DEBUG
+logging.level.org.apache.isis.applib.services.queryresultscache.QueryResultsCache = DEBUG
 logging.level.org.apache.isis.core.runtime.persistence.transaction.ChangedObjectsService = DEBUG
